+++ library/src/main/java/blog/softwaretester/properties/propertysource/EnvironmentPropertiesSource.java
@@ -1,12 +1,15 @@
 package blog.softwaretester.properties.propertysource;
 
+import blog.softwaretester.properties.PropertyConverter;
+
+import java.util.Map;
 import java.util.Properties;
 
 public final class EnvironmentPropertiesSource implements PropertySource {
     @Override
-    public Properties getProperties() {
+    public Map<String, String> getProperties() {
         Properties properties = new Properties();
         properties.putAll(System.getenv());
-        return properties;
+        return PropertyConverter.propertiesToMap(properties);
     }
 }