+++ library/src/main/java/blog/softwaretester/properties/propertysource/PropertySource.java
@@ -1,16 +1,16 @@
 package blog.softwaretester.properties.propertysource;
 
-import java.util.Properties;
+import java.util.Map;
 
 /**
  * The interface for all property sources.
  */
 public interface PropertySource {
     /**
-     * Get a {@link Properties} instance containing all properties from the
+     * Get a string map containing all properties from the
      * respective {@link PropertySource}.
      *
-     * @return The {@link Properties} instance.
+     * @return The string map.
      */
-    Properties getProperties();
+    Map<String, String> getProperties();
 }