+++ CHANGELOG.md
@@ -10,10 +10,17 @@ and this project adheres to
 
 Back to [Readme](README.md).
 
+[1.3.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/bischoffdev/property-aggregator/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/bischoffdev/property-aggregator/tree/v1.0.0
+
+## [1.3.0] - UNRELEASED
+
+### Changed
+* Use commons logging for compatibility
+* Changed internal property storage to String map
 
 ## [1.2.1] - 2022-03-14